requirements.txt
Django==4.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
# django-model-utils==4.3.1
xlwt=1.3.0
xlutils=2.0.0
xlrd=2.0.1
django-cors-headers=3.13.0